requirements.txt
bokeh==3.6.1
chembl_structure_pipeline==1.2.2
ipython==8.12.3
joblib==1.4.2
matplotlib==3.9.4
mordredcommunity[full]==2.0.6
numpy==1.26.4
openpyxl==3.1.5
pandas==2.2.3
Pillow==11.0.0
py3Dmol==2.4.2
pytest==8.3.4
rdkit==2024.9.6
scikit_learn<=1.6.0
scipy==1.15.3
seaborn==0.13.2
selfies==2.1.2
shap==0.46.0
umap-learn==0.5.7
